# openapi/bard-calendar.openapi.yaml
openapi: 3.1.0
info:
  title: Bard Calendar Agent API
  version: 0.2.0
  description: >-
    Authenticated API for managing content Topics, channel Drafts, publishing
    Events, Prism bundle runs, and generated asset references.
servers:
  - url: /
    description: Current Bard Calendar origin.
...
security:
  - bearerAuth: []
paths:
...
  /api/agent/topics/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getTopic
      tags: [Topics]
      summary: Get a Topic
      responses:
        '200':
          description: Topic found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateTopic
      tags: [Topics]
      summary: Update a Topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicPatch'
      responses:
        '200':
          description: Topic updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteTopic
      tags: [Topics]
      summary: Delete a Topic and its Drafts
      responses:
        '200':
          description: Topic deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/agent/drafts/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getDraft
      tags: [Drafts]
      summary: Get a Draft
      responses:
        '200': { description: Draft found }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateDraft
      tags: [Drafts]
      summary: Update a Draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftPatch'
      responses:
        '200': { description: Draft updated }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteDraft
      tags: [Drafts]
      summary: Delete a Draft
      responses:
        '200': { description: Draft deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/agent/bundle-runs/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getBundleRun
      tags: [Bundles]
      summary: Get a content bundle run
      responses:
        '200':
          description: Bundle run found
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run:
                    $ref: '#/components/schemas/BundleRun'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateBundleRun
      tags: [Bundles]
      summary: Update a Prism content bundle run
      description: >-
        Prism uses this callback to report source revision, progress, audit
        results, terminal state, and failures for the run ID in its hook payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BundleRunUpdate'
      responses:
        '200':
          description: Bundle run updated
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run:
                    $ref: '#/components/schemas/BundleRun'
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/agent/content-assets/upsert:
    put:
      operationId: upsertContentAsset
      tags: [Bundles]
      summary: Upsert a Prism-generated content asset reference
      description: Idempotent by prism_request_id and prism_artifact_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentAssetUpsert'
      responses:
        '200': { description: Asset reference upserted }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Prism request, Topic, or Draft relationship conflict
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidRequest:
      description: Invalid request body or query
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: { type: object, additionalProperties: true }
...
    Metadata:
      type: object
      additionalProperties: true
    TopicStatus:
      type: string
      enum: [active, archived]
    DraftStatus:
      type: string
      enum: [draft, ready, assigned, published, archived]
...
    TopicPatch:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        supporting_material_markdown: { type: [string, 'null'] }
        status:
          $ref: '#/components/schemas/TopicStatus'
        created_by: { type: [string, 'null'] }
        metadata:
          $ref: '#/components/schemas/Metadata'
        external_source: { type: [string, 'null'] }
        external_id: { type: [string, 'null'] }
...
    DraftPatch:
      type: object
      properties:
        topic_id: { type: string }
        title: { type: string, minLength: 1 }
        target_channel: { type: string, minLength: 1 }
        markdown_content: { type: string }
        external_draft_url: { type: [string, 'null'], format: uri }
        status:
          $ref: '#/components/schemas/DraftStatus'
        created_by: { type: [string, 'null'] }
        metadata:
          $ref: '#/components/schemas/Metadata'
        external_source: { type: [string, 'null'] }
        external_id: { type: [string, 'null'] }
...
    AuditCheck:
      type: object
      required: [key, label, status]
      properties:
        key: { type: string }
        label: { type: string }
        status:
          type: string
          enum: [pass, warning, fail, not_checked]
        evidence: { type: [string, 'null'] }
    BundleRunUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [queued, running, partial, complete, failed, canceled]
        stage: { type: string, minLength: 1 }
        source_revision: { type: [string, 'null'] }
        prism_request_id: { type: [string, 'null'] }
        audit_checks:
          type: array
          items:
            $ref: '#/components/schemas/AuditCheck'
        error_message: { type: [string, 'null'] }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
    BundleRun:
      allOf:
        - $ref: '#/components/schemas/BundleRunUpdate'
        - type: object
          required:
            - id
            - topic_id
            - source_system
            - source_id
            - requested_channels
            - options
            - created_at
            - updated_at
          properties:
            id: { type: string }
            topic_id: { type: string }
            source_system: { type: string }
            source_id: { type: string }
            source_url: { type: [string, 'null'], format: uri }
            requested_channels:
              type: array
              items: { type: string }
            options:
              $ref: '#/components/schemas/Metadata'
            instructions: { type: [string, 'null'] }
            created_by: { type: [string, 'null'] }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    ContentAssetUpsert:
      type: object
      required: [topic_id, kind, prism_request_id, prism_artifact_id]
      properties:
        topic_id: { type: string }
        draft_id: { type: [string, 'null'] }
        kind: { type: string, minLength: 1 }
        status:
          type: string
          enum: [generated, approved, rejected, archived]
        target_channel: { type: [string, 'null'] }
        prism_request_id: { type: string, minLength: 1 }
        prism_artifact_id: { type: string, minLength: 1 }
        stable_url: { type: [string, 'null'], format: uri }
        mime_type: { type: [string, 'null'] }
        prompt: { type: [string, 'null'] }
        metadata:
          $ref: '#/components/schemas/Metadata'